test_drf/requirements.txt
django==3.2
uwsgi==2.0.20
flake8==4.0
django-debug-toolbar==3.4.0
django-split-settings
django-dotenv
django-filter
pyyaml
uritemplate
psycopg2